VTK: Moved definition of outputlogger to header file and made it a private inner class, can't forward declare when using vtkSmartPointer
